Improved bitrate re-calculation.
Now, when the end file size is above target, the audio bitrate (if it
wasn't previously available) will be checked and subtracted from the
video bitrate. Otherwise (if audio bitrate is still unobtainable or
there has already been a previous transcoding attempt) the factor is
further reduced.

All this is to reduce the number of transcoding attempts the program
makes when the output file size is larger than target. As there is no
tolerance of the end file size being even slightly above target, the
script would previously take a *long* time (30+ attempts) to reduce the
file size in small increments. This is because it was previously trying
to get the bitrate calculation *exactly* on target, rather than a bit
below it. Now, it manages to meet the target (within 10% tolerance)
within 1-3 attempts with my own testing.